+++ src/main/java/com/minecolonies/coremod/colony/buildings/modules/CourierAssignmentModule.java
@@ -8,7 +8,6 @@ import com.minecolonies.api.colony.jobs.ModJobs;
 import com.minecolonies.api.colony.jobs.registry.JobEntry;
 import com.minecolonies.coremod.colony.jobs.JobDeliveryman;
 import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.network.PacketBuffer;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
@@ -20,16 +19,11 @@ import static com.minecolonies.api.util.constant.NbtTagConstants.*;
  */
 public class CourierAssignmentModule extends AbstractAssignedCitizenModule implements IAssignsJob, IBuildingEventsModule, ITickingModule, IPersistentModule
 {
-    /**
-     * The hiring mode of this particular building, by default overriden by colony mode.
-     */
-    private HiringMode hiringMode = HiringMode.DEFAULT;
-
     @Override
     public void onColonyTick(@NotNull final IColony colony)
     {
         // If we have no active worker, grab one from the Colony
-        if (!isFull() && (this.hiringMode == HiringMode.DEFAULT && !building.getColony().isManualHiring() || this.hiringMode == HiringMode.AUTO))
+        if (!isFull() && (this.getHiringMode() == HiringMode.DEFAULT && !building.getColony().isManualHiring() || this.getHiringMode() == HiringMode.AUTO))
         {
             for (final ICitizenData data : colony.getCitizenManager().getCitizens())
             {
@@ -53,7 +47,6 @@ public class CourierAssignmentModule extends AbstractAssignedCitizenModule imple
     public void deserializeNBT(final CompoundNBT compound)
     {
         final CompoundNBT jobCompound = compound.getCompound("warehouse");
-        this.hiringMode = HiringMode.values()[jobCompound.getInt(TAG_HIRING_MODE)];
         final int[] residentIds = jobCompound.getIntArray(TAG_COURIERS);
         for (final int citizenId : residentIds)
         {
@@ -69,7 +62,6 @@ public class CourierAssignmentModule extends AbstractAssignedCitizenModule imple
     public void serializeNBT(final CompoundNBT compound)
     {
         final CompoundNBT jobCompound = new CompoundNBT();
-        jobCompound.putInt(TAG_HIRING_MODE, this.hiringMode.ordinal());
         if (!assignedCitizen.isEmpty())
         {
             final int[] residentIds = new int[assignedCitizen.size()];
@@ -80,13 +72,6 @@ public class CourierAssignmentModule extends AbstractAssignedCitizenModule imple
             jobCompound.putIntArray(TAG_COURIERS, residentIds);
         }
         compound.put("warehouse", jobCompound);
-    }
-
-    @Override
-    public void serializeToView(final @NotNull PacketBuffer buf)
-    {
-        super.serializeToView(buf);
-        buf.writeInt(hiringMode.ordinal());
     }
 
     @Override
@@ -111,16 +96,5 @@ public class CourierAssignmentModule extends AbstractAssignedCitizenModule imple
     public JobEntry getJobEntry()
     {
         return ModJobs.delivery;
-    }
-
-    public void setHiringMode(final HiringMode hiringMode)
-    {
-        this.hiringMode = hiringMode;
-        this.markDirty();
-    }
-
-    public HiringMode getHiringMode()
-    {
-        return hiringMode;
     }
 }
