+++ src/main/java/com/minecolonies/coremod/colony/buildings/modules/WorkerBuildingModule.java
@@ -58,11 +58,6 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
      */
     private final Function<IBuilding, Integer> sizeLimit;
 
-    /**
-     * The hiring mode of this particular building, by default overriden by colony mode.
-     */
-    private HiringMode hiringMode = HiringMode.DEFAULT;
-
     public WorkerBuildingModule(
       final JobEntry entry,
       final Skill primary,
@@ -113,7 +108,6 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
                     assignCitizen(data);
                 }
             }
-            this.hiringMode = HiringMode.values()[compound.getInt(TAG_HIRING_MODE)];
         }
         else if (compound.contains(jobEntry.getKey().toString()))
         {
@@ -127,7 +121,6 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
                     assignCitizen(citizen);
                 }
             }
-            this.hiringMode = HiringMode.values()[jobCompound.getInt(TAG_HIRING_MODE)];
         }
     }
 
@@ -136,7 +129,7 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
     {
         // If we have no active worker, grab one from the Colony
         if (!isFull() && ((building.getBuildingLevel() > 0 && building.isBuilt()) || building instanceof BuildingBuilder)
-              && (this.hiringMode == HiringMode.DEFAULT && !building.getColony().isManualHiring() || this.hiringMode == HiringMode.AUTO))
+              && (this.getHiringMode() == HiringMode.DEFAULT && !building.getColony().isManualHiring() || this.getHiringMode() == HiringMode.AUTO))
         {
             final ICitizenData joblessCitizen = colony.getCitizenManager().getJoblessCitizen();
             if (joblessCitizen != null)
@@ -152,14 +145,13 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
         final CompoundNBT jobCompound = new CompoundNBT();
         if (!assignedCitizen.isEmpty())
         {
-            @NotNull final int[] residentIds = new int[assignedCitizen.size()];
+            final int[] residentIds = new int[assignedCitizen.size()];
             for (int i = 0; i < assignedCitizen.size(); ++i)
             {
                 residentIds[i] = assignedCitizen.get(i).getId();
             }
             jobCompound.putIntArray(TAG_WORKING_RESIDENTS, residentIds);
         }
-        jobCompound.putInt(TAG_HIRING_MODE, this.hiringMode.ordinal());
         compound.put(jobEntry.getKey().toString(), jobCompound);
     }
 
@@ -168,7 +160,6 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
     {
         super.serializeToView(buf);
         buf.writeRegistryId(jobEntry);
-        buf.writeInt(hiringMode.ordinal());
         buf.writeInt(getModuleMax());
         buf.writeInt(getPrimarySkill().ordinal());
         buf.writeInt(getSecondarySkill().ordinal());
@@ -228,19 +219,6 @@ public class WorkerBuildingModule extends AbstractAssignedCitizenModule implemen
     public IJob<?> createJob(final ICitizenData citizen)
     {
         return jobEntry.produceJob(citizen);
-    }
-
-    @Override
-    public void setHiringMode(final HiringMode hiringMode)
-    {
-        this.hiringMode = hiringMode;
-        this.markDirty();
-    }
-
-    @Override
-    public HiringMode getHiringMode()
-    {
-        return hiringMode;
     }
 
     @Override
